lat, lon double 자료형으로 변경

+++ insider/src/main/java/com/kh/insider/dto/MemberDto.java
@@ -17,8 +17,8 @@ public class MemberDto {
 	private String memberEmail;
 	 @DateTimeFormat(pattern = "E MMM dd HH:mm:ss z yyyy")
 	private Date memberLogin;
-	private int memberLat;
-	private int memberLon;
+	private double memberLat;
+	private double memberLon;
 	private String memberPost;
 	private String memberBasicAddr;
 	private String memberDetailAddr;
